Renamed to 'definitions'
